requirements.txt
psycopg2-binary
google-cloud-bigquery
requests
pyyaml
